+++ asdf/_tests/test_asdf.py
@@ -5,10 +5,10 @@ import pytest
 
 from asdf import config_context, get_config
 from asdf._tests._helpers import assert_no_warnings, assert_tree_match, yaml_to_asdf
-from asdf.asdf import AsdfFile, SerializationContext, open_asdf
+from asdf.asdf import AsdfFile, open_asdf
 from asdf.entry_points import get_extensions
 from asdf.exceptions import AsdfWarning
-from asdf.extension import ExtensionManager, ExtensionProxy
+from asdf.extension import ExtensionManager, ExtensionProxy, SerializationContext
 from asdf.extension._legacy import AsdfExtensionList
 from asdf.versioning import AsdfVersion
