+++ asdf/extension/__init__.py
@@ -9,6 +9,7 @@ from ._converter import Converter, ConverterProxy
 from ._extension import Extension, ExtensionProxy
 from ._manager import ExtensionManager, get_cached_extension_manager
 from ._manifest import ManifestExtension
+from ._serialization_context import SerializationContext
 from ._tag import TagDefinition
 from ._validator import Validator
 
@@ -24,4 +25,5 @@ __all__ = [
     "ConverterProxy",
     "Compressor",
     "Validator",
+    "SerializationContext",
 ]
